requirements.txt
python-telegram-bot==21.1
Flask>=3.0.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
dateparser==1.1.8
pytest==8.0.0
pytest-asyncio==0.23.0
